requirements.txt
fastapi
uvicorn
streamlit
requests
sqlalchemy
passlib[bcrypt]              
python-multipart
jinja2 
python-dotenv